+++ src/ftxui/screen/screen.cpp
@@ -14,7 +14,7 @@
 #ifndef NOMINMAX
 #define NOMINMAX
 #endif
-#include <Windows.h>
+#include <windows.h>
 #endif
 
 namespace ftxui {